+++ js/src/elements/tab-container/index.ts
@@ -1,9 +1,9 @@
 // This component is a copy, rename and extension of the library
 // https://github.com/github/tab-container-element/releases/tag/v4.8.0
 
-import { TabContainerElement } from './segmented-controls-element';
+import { TabContainerElement } from './element';
 
 export { TabContainerElement };
 export default TabContainerElement;
 
-export * from './segmented-controls-element-define';
+export * from './define';
