+++ js/src/index.ts
@@ -2,5 +2,5 @@ export * from './behaviors/dismiss';
 export * from './behaviors/popover';
 export * from './behaviors/snap-scroll';
 export * from './elements/form-behavior';
-export * from './elements/segmented-controls/index';
+export * from './elements/tab-container/index';
 export * from './utilities/util';
